jv - implement rating system scoring algos

+++ server/src/main/java/com/github/animelist/animelist/model/userlist/UserListRating.java
@@ -7,11 +7,11 @@ public class UserListRating {
 
     private String displayRating;
 
-    private Integer rating;
+    private Double rating;
 
     private List<UserListSubRating> subRatings;
 
-    public UserListRating(String displayRating, Integer rating, List<UserListSubRating> subRatings) {
+    public UserListRating(String displayRating, Double rating, List<UserListSubRating> subRatings) {
         this.displayRating = displayRating;
         this.rating = rating;
         this.subRatings = subRatings;
@@ -38,11 +38,11 @@ public class UserListRating {
         this.displayRating = displayRating;
     }
 
-    public Integer getRating() {
+    public Double getRating() {
         return rating;
     }
 
-    public void setRating(Integer rating) {
+    public void setRating(Double rating) {
         this.rating = rating;
     }
 
@@ -60,7 +60,7 @@ public class UserListRating {
 
     public static class Builder {
         private String displayRating;
-        private Integer rating;
+        private Double rating;
         private List<UserListSubRating> subRatings;
 
         public Builder displayRating(String displayRating) {
@@ -68,7 +68,7 @@ public class UserListRating {
             return this;
         }
 
-        public Builder rating(Integer rating) {
+        public Builder rating(double rating) {
             this.rating = rating;
             return this;
         }
